Add 'lit-page' to package-lock.json dependencies
The ‘lit-page’ component was inserted into the project dependencies in 'package-lock.json'. This component is required for new page functionality. The dependency has been added under both components and node_modules sections to ensure the correct mapping and reference to the 'lit-page' component. A version has also been allocated for better version control and project maintenance.